Blind compilation fix.
git-svn-id: http://svn.code.sf.net/p/wxperl/code/wxPerl/trunk@2522 2e9104e8-d5c8-4c90-bb67-bfd1792b7337

+++ cpp/helpers.cpp
@@ -5,7 +5,7 @@
 // Modified by:
 // Created:     29/10/2000
 // RCS-ID:      $Id$
-// Copyright:   (c) 2000-2008 Mattia Barbon
+// Copyright:   (c) 2000-2009 Mattia Barbon
 // Licence:     This program is free software; you can redistribute it and/or
 //              modify it under the same terms as Perl itself
 /////////////////////////////////////////////////////////////////////////////
@@ -1287,7 +1287,7 @@ wxVariant wxPli_sv_2_wxvariant( pTHX_ SV* sv )
         // TODO
         return wxVariant();
     } else if( SvNOK( sv ) ) {
-        return wxVariant( SvNV( sv ) );
+        return wxVariant( (double)SvNV( sv ) );
     } else if( SvIOK( sv ) ) {
 #if INTSIZE > LONGSIZE
         return wxVariant( (int)SvIV( sv ) );
